requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiohttp>=3.9.0
geopy>=2.4.0
